ML-Based Stroke Risk Prediction
The primary objective of this project is to develop a predictive model that can accurately determine the presence of heart disease in patients based on a set of medical attributes. This model aims to assist in early diagnosis and potential intervention, improving patient outcomes and healthcare efficiency.

The project involves analyzing the Heart Disease dataset from the UCI Machine Learning Repository using Python and Jupyter Notebook. The analysis will leverage libraries such as numpy and pandas for data manipulation, and sklearn.model_selection for splitting the dataset into training and test sets. Logistic Regression will be used to build a predictive model to determine the presence of heart disease based on various medical attributes.